Move interim references rewriting (to fix validation issues) to validate.py, and a simple automated test for validate to start with.
