+++ src/validate.py
@@ -1,4 +1,4 @@
-import os, sys, json
+import os, sys, json, re
 import conf
 import jsonschema
 
@@ -230,6 +230,36 @@ def wrap_body_in_section(body_json):
     # Continue with rewriting
     return body_json
 
+def references_rewrite(references):
+    "clean up values that will not pass validation temporarily"
+    for ref in references:
+        if "date" in ref:
+            # Scrub non-numeric values from the date, which comes from the reference year
+            ref["date"] = re.sub("[^0-9]", "", ref["date"])
+        if ref.get("type") == "other":
+            # The schema cannot support type other, turn this into a basic journal reference
+            #  to pass validation
+            ref["type"] = "journal"
+            #if not "articleTitle" in ref:
+            #    ref["articleTitle"] = "Placeholder article title for ref of type 'other'"
+            if not "journal" in ref:
+                ref["journal"] = {}
+                ref["journal"]["name"] = []
+                #ref["journal"]["name"].append("This is a transformed placeholder journal name for ref of type 'other'")
+                if "source" in ref:
+                    ref["journal"]["name"].append(ref["source"])
+                    del ref["source"]
+        if ref.get("type") == "journal" and not "pages" in ref:
+            #ref["pages"] = "placeholderforrefwithnopages"
+            pass
+        if ref.get("type") == "book":
+            if not "publisher" in ref:
+                ref["publisher"] = {}
+                ref["publisher"]["name"] = []
+                #ref["publisher"]["name"].append("This is a placeholder book publisher name for ref that does not have one")
+
+    return references
+
 def is_poa(contents):
     try:
         return contents["article"]["status"] == "poa"
@@ -248,7 +278,7 @@ def add_placeholders_for_validation(contents):
         del art['relatedArticles']
 
     if 'references' in art:
-        art['references'] = placeholder_references
+        art['references'] = references_rewrite(placeholder_references)
 
     if 'published' not in art:
         art['published'] = '1970-07-01T00:00:00Z'
